chore: move BridgeToken type to client
